move extension for limit characters for flight number and check number of characters of flight number to a ViewModel
